pyproject.toml
[project]
name = "mindclaw"
version = "0.1.0"
description = "Personal AI assistant framework - secure, extensible, multi-channel, multi-model"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.55",
    "pydantic>=2.0",
    "typer>=0.15",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "httpx>=0.28",
    "loguru>=0.7",
    "websockets>=14.0",
    "python-telegram-bot>=21.0",
    "cryptography>=44.0",
    "croniter>=1.4",
    "slack-sdk>=3.33",
    "markdown-to-mrkdwn>=0.3.1",
]

[project.optional-dependencies]
slack = ["slack-sdk>=3.33"]
feishu = ["lark-oapi>=1.0"]
discord = ["discord.py>=2.0"]
all-channels = [
    "slack-sdk>=3.33",
    "lark-oapi>=1.0",
    "discord.py>=2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
]
vector = [
    "lancedb>=0.6",
    "pyarrow>=15.0",
]

[project.scripts]
mindclaw = "mindclaw.cli.commands:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
